Fixed c plus plus sanitizer to account for File Content or Code
closes https://github.com/SoftUni-Internal/suls-issues/issues/5671
